requirements.txt
langchain 
langchain-google-genai 
langchain-groq 
chromadb 
python-dotenv 
pandas 
PyPDF2 
python-docx 
openpyxl 
moviepy==1.0.3
requests 
crawl4ai
aiohttp
langchain-community
langchain-core
pypdf
fastapi[all]
uvicorn
unstructured
streamlit